Removed IntermediaryFilesHelper
