+++ rabix-engine/src/main/java/org/rabix/engine/service/IntermediaryFilesHandler.java
@@ -1,0 +1,10 @@
+package org.rabix.engine.service;
+
+import java.util.Set;
+
+import org.rabix.bindings.model.Job;
+
+public interface IntermediaryFilesHandler {
+
+ public void handleUnusedFiles(Job job, Set<String> unusedFiles);
+}
