+++ rabix-engine/src/main/java/org/rabix/engine/service/impl/IntermediaryFilesServiceLocalHandler.java
@@ -7,26 +7,24 @@ import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Set;
-import java.util.UUID;
 
-import org.rabix.engine.service.IntermediaryFilesService;
+import org.rabix.bindings.model.Job;
+import org.rabix.engine.service.IntermediaryFilesHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.inject.Inject;
 
-public class IntermediaryFilesServiceLocalImpl extends IntermediaryFilesService {
+public class IntermediaryFilesServiceLocalHandler implements IntermediaryFilesHandler {
   
-  private final static Logger logger = LoggerFactory.getLogger(IntermediaryFilesServiceLocalImpl.class);
-
+  private final static Logger logger = LoggerFactory.getLogger(IntermediaryFilesServiceLocalHandler.class);
+  
   @Inject
-  public IntermediaryFilesServiceLocalImpl () {
-    super();
+  private IntermediaryFilesServiceLocalHandler() {
   }
   
   @Override
-  public void handleUnusedFiles(UUID rootId) {
-    Set<String> unusedFiles = getUnusedFiles(rootId);
+  public void handleUnusedFiles(Job job, Set<String> unusedFiles) {
     for (String p : unusedFiles) {
       Path path = Paths.get(p);
       try {
